UTIL-352 Quantify to be int value

+++ production/src/main/java/com/herbmarshall/base/diff/DiffGenerator.java
@@ -27,12 +27,13 @@ public interface DiffGenerator {
 
 	/**
 	 * Quantify the difference between two objects.
+	 * Value will be in 1/10000 units.  So 420 = 4.20%
 	 * @param expected The expected value to compare with.
 	 * @param actual The actual value to compare to.
 	 * @return The quantified difference between the expected and actual values.
 	 */
 	// TODO Remove this default on 2.0 <a href="https://herbmarshall.atlassian.net/browse/UTIL-353">UTIL-353</a>.
-	default double quantify( Object expected, Object actual ) {
+	default int quantify( Object expected, Object actual ) {
 		// TODO Remove default implementation
 		return DiffGeneratorDefault.INSTANCE.quantify( expected, actual );
 	}
